pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "nltk>=3.9.1",
    "pandas>=2.3.0",
    "pdfplumber>=0.11.7",
    "pillow>=11.2.1",
    "plotly>=6.1.2",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "scikit-learn>=1.7.0",
    "spacy>=3.8.7",
    "streamlit>=1.46.0",
    "textstat>=0.7.7",
    "yake>=0.4.8",
]
